fix guild filter
