+++ src/components/patreon/PatreonSettingsEditor.tsx
@@ -1,14 +1,11 @@
 import {
 	FC,
-	useMemo,
 	useState
 }                           from "react";
 import { useParams }        from "react-router-dom";
 import useGuild             from "@hooks/useGuild";
 import {
 	channelToSelectedSingle,
-	channelToSelection,
-	rolesToSelection,
 	roleToSelectedMulti
 }                           from "@lib/selectConversion";
 import {
@@ -24,18 +21,16 @@ import {
 	tRPC_handleError
 }                           from "@lib/tRPC";
 import { fireToastFromApi } from "@lib/sweetAlert";
+import useSelection         from "@hooks/useSelection";
 
 
 const PatreonSettingsEditor : FC = () => {
+	const { forumChannelOptions, textChannelOptions, roleOptions } = useSelection();
 	const { guildId } = useParams();
 	const [ isLoading, setIsLoading ] = useState( false );
 
 	const { textChannels, forumChannels, roles, guildData, triggerGuildUpdate } = useGuild();
 	const { patreonOptions } = guildData;
-
-	const roleOptions = useMemo( () => rolesToSelection( roles ), [ roles ] );
-	const textChannelOptions = useMemo( () => channelToSelection( textChannels ), [ textChannels ] );
-	const forumChannelOptions = useMemo( () => channelToSelection( forumChannels ), [ forumChannels ] );
 
 	const [ announcementChannel, setAnnouncementChannel ] = useState( () => channelToSelectedSingle( textChannels, patreonOptions?.announcementChannel || "" ) );
 	const [ changelogForum, setChangelogForum ] = useState( () => channelToSelectedSingle( forumChannels, patreonOptions?.changelogForum || "" ) );
